Convert edges to superedges for boundary and advection
This converts all edge-loops (both domain-, and boundary) to use
superedges for both the gradient calculation and advection. The
superedges are commented out for now, because enabling them will require
some minor adjustment of some of the regression test baselines so this
change is separate from that. (This commit should pass all tests.)