Move building Charm++ to the xyst build
This removes the need for two build stages: no more separate TPL build.

Building is now as simple as

cmake <src-dir> && make

Charm++ is pulled in via cmake's external project functionality as
before and cmake code is written to ensure that

* Charm++ is not part of the 'all' target but every object that depends
  on it triggers its build,

* 'make clean' will not remove Charm++'s install from the build dir,

* as a result, Charm++ is only built for the first time and will not be
  rebuilt during incremental builds even after 'make clean'.

The above makes it pretty usable and now the build process is single
stage, which simplifies the build significantly.

HDF5 and NetCDF are assumed available; will not maintain builds for
those.