Add AMPI-based build
This allows building on Charm++'s AMPI instead of MPI.

It requires to build HDF5 (a patched version) and NetCDF. Builds fine
but I'm missing something because it does not appear to run
Transporter's constructor. So needs some more work/understanding but
keeping the code in the hope this will be useful eventually.

Also changes is how we build Charm++: Instead of the INTERFACE stuff in
cmake, now if charm is not found we write a separate CMakeLists.txt and
call cmake on it within the build dir. This allows using cmake's
externalproject to be used together with find_package. The previous
solution would not work if libraries need to be found.

Also simplified some cmake code.

Also remove version.

Also remove exception testing from mpi.