Re-enable Zoltan's hypergraph partitioner
Unfortunately, this does not work with distributed meshes, because
Zoltan does not see the complete graph on input.

If "CHECK_HYPERGRAPH", "1" is set, it produces warnings like

phg_hypergraph.c):  Found multiple vertex for same hedge.

and

phg.c:  Error building hypergraph.
Zoltan ERROR in Zoltan_PHG (line 506 of phg.c):  Error partitioning hypergraph.

A possible solution might be to perform some communication before
passing the graph to Zoltan - so keep this code for the future.