Add back Zoltan2
Note: Build own NetCDF as well? Not a bad idea as it would eliminate the
need to install it system-wide. However, since exodus needs NetCDF4
which needs HDF5, building our own HDF5 while would be possible, it
would lead to maintenance headaches because the netcdf build command
needs to be told where HDF5 is and this is tricky and error-prone
depending whether hdf5 was built by us or found on the system. Better
have the user/developer load netcdf and hdf5 that comes with the system
or build by e.g., spack.