Replace 'skip' with 'ign'
Apparently 'skip' is an 'action' and not a 'constraint' in numdiff and
they way this was used may yield a rule/constraint that causes to ignore
everything in the file, leading to silently accept every field as pass.
The correct way to ignore values is to use 'ign'. This commit fixes one
numdiff config file and also updates the baseline which now would
correctly fail based on the constraints specified and not ignored.