Add Neumann BCs to conjugate gradients solve
Based on a new test case that solves the Poisson equation with an
analytic solution, add code to set inhomogeneous Neumann BCs to the
conjugate gradients solve. In parallel this also required some
communication code, piggy-backed on the already existing communication
step for the BC data within the ConjugateGradients class. Also augmented
the and clarified the verbose outputs within the conjugate gradients
solve and the exit condition logic.