Trigger error if input exo mesh contains elem blocks other than TRI or TET
If the input exodus mesh contains element block types other than
triangles and tetrahedra, the reader could silently erroneously read
through the file, producing boundary elements and/or nodes associated to
incorrect side sets, screwing up boundary conditions.

This commit puts in code to trigger an error if an input mesh with
element blocks other than triangles and/or tetrahedra is supplied by the
user. A regressopm test is also added to test that the error is indeed
triggered for such a case.