Add absolute performance metric to status
This is called estimated grind performance, EGP, and is based on
ms/dof/timestep to enable absolution performance comparison of different
solvers. This is computed as EGT/DOF, where EGT is the estimated grind
time (the wall-clock time it takes to take a time step) and DOF is the
total number of points (across all meshes).

Using the same mesh on the same number of CPUs this enables direct
comparison of two solvers, e.g., ChoCG and LohCG, based on grind time,
i.e., the wall-clock time it takes to take a single time step.