Use non-reconstructed pressure in LohCG::damp4
Measured with computing Poiseuille flow, this does not reduce the
absolute error and does not change the convergece rate, but requires no
pressure gradients and so it is faster and uses less memory. It is also
more stable and allows larger CFL / time steps.

The measured convergence rates are damp4+rk4:p=2.5 and damp2+rk3:p=2.0.