Fix srun ncpus arg
This commit fixes how the ncpus arg is assumed to be interpreted if srun
is used as the test runner.

Previously it was assumed that independent of the RUNNER (e.g., mpirun,
srun) the RUNNER_NCPUS_ARG meant the total number of PEs the test uses.
This is correct in Charm++'s non-SMP mode, however, incorrect in SMP
mode, since mpirun -n NPE ... +ppn 2 uses a total of NPE cores, but srun
-n NPE ... +ppn 2 uses a total of NPE * 2 cores.

The test_runner now takes this into account and thus correct in SMP
mode. Non-SMP mode is not affected.

Also removed the user/developer ability to configure RUNNER_NCPUS_ARG,
instead, it is always -n if mpirun or srun is given as the RUNNER and
stays +p for charmrun as the default. This is less error-prone and
easier to use.