Enable using different partitioners for different meshes
To make the logic correct in this case, it is now required to finish
loading all meshes before even first mesh can be partitioned.

This is because Transporter::load() computes the load distribution and
before this commit it has been possible that execution reached
Transporter::partitioned(), spawning partitioning the second mesh before
Transporter::load() was called for it. This was incorrect, because the
load distribution at that point was not yet computed for the second
mesh. This commit fixes this problem by requiring the load of all meshes
before any of the meshes can be partitioned.