Put in error if more than 2 meshes are given
Overset will not need to support more than 2 meshes: the first one is
assumed to be the background mesh and the second one will be the
embedded or overset mesh. If more overset meshes are necessary, they can
be combined into a single exodus file. As a result, there is no need to
support mesh-to-mesh solution transfer between more than a pair of
meshes.