Overset (to and from) appears to work
Serial, parallel, with and without virtualization, with unqueal
virtualization, one-way or two-way transfer.

The key change required was the introduction of a callback to the source
side besdides the existing destination side, called, once per source
chare, from detemineActuallCollisions().

Tested with transferring a linear scalar field, using LohCG, only in
one-way for now, in both serial and parallel + unequal virtualization.

Not yet tested in SMP mode.