User freezeflow only for variable dt
This commit makes freezeflow ignored in ChoCG if the user wants constant
dt. Constant dt is a very specific case and intended only for testing,
for which cases, it is arguably important NOT to interfere with the
user's (developer's) setting. This does not affect baselines (because a
combination of const dt + freezeflow is not tested).